move snapkv util to kv
